pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bcf-governance"
version = "0.1.0"
description = "Executable governance templates and validators for agent-led development."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "BCF Governance Maintainers" }]
dependencies = [
  "PyYAML>=6.0,<7",
  "jsonschema>=4.21,<5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9"]

[project.scripts]
bcf = "bcf_governance.cli:main"

[tool.setuptools.packages.find]
include = ["bcf_governance*", "scripts"]

[tool.setuptools.package-data]
bcf_governance = ["pack/template-repo/**/*"]

[tool.setuptools.exclude-package-data]
bcf_governance = [
  "pack/template-repo/**/__pycache__/*",
  "pack/template-repo/**/*.pyc",
]
